Add basic bot setup
This commit adds basic support for running a bot that will take in a
yaml configuration which describes valid elections, listen to imap for
nominations respond to them when valid and respond to valid nominations.

It also has some basic wiki page handling support which is very
incomplete at this point. It still needs support for updating the wiki
when a valid candidate nomination is added.

The other major pieces missing are support for pulling the list of the
electorate and starting the election on CIVS, checking the candidate
is a eligable to run, and probably a few things I can't remember.

It's also worth pointing out nothing at this point is tested so this
is very much just code to back up an idea.